MAINT: integrate: break dependency on optimize
