+++ scipy/integrate/_ivp/ivp.py
@@ -4,7 +4,7 @@ from .bdf import BDF
 from .radau import Radau
 from .rk import RK23, RK45, DOP853
 from .lsoda import LSODA
-from scipy.optimize import OptimizeResult
+from scipy._lib._util import OptimizeResult
 from .common import EPS, OdeSolution
 from .base import OdeSolver
