+++ scipy/optimize/_cobyla_py.py
@@ -14,9 +14,9 @@ import functools
 from threading import RLock
 
 import numpy as np
+from scipy._lib._util import OptimizeResult
 from scipy.optimize import _cobyla as cobyla
-from ._optimize import (OptimizeResult, _check_unknown_options,
-    _prepare_scalar_function)
+from ._optimize import _check_unknown_options, _prepare_scalar_function
 try:
     from itertools import izip
 except ImportError:
