+++ scipy/optimize/_differentialevolution.py
@@ -6,12 +6,12 @@ import warnings
 
 import numpy as np
 
-from scipy.optimize import OptimizeResult, minimize
+from scipy.optimize import minimize
 from scipy.optimize._constraints import (Bounds, new_bounds_to_old,
                                          NonlinearConstraint, LinearConstraint)
 from scipy.optimize._optimize import _status_message, _wrap_callback
-from scipy._lib._util import (check_random_state, MapWrapper, _FunctionWrapper,
-                              rng_integers, _transition_to_rng)
+from scipy._lib._util import (OptimizeResult, check_random_state, MapWrapper,
+                              _FunctionWrapper, rng_integers, _transition_to_rng)
 from scipy._lib._sparse import issparse
