+++ scipy/optimize/_lbfgsb_py.py
@@ -35,8 +35,10 @@ Functions
 
 import numpy as np
 from numpy import array, asarray, float64, zeros
+
+from scipy._lib._util import OptimizeResult
 from . import _lbfgsb
-from ._optimize import (MemoizeJac, OptimizeResult, _call_callback_maybe_halt,
+from ._optimize import (MemoizeJac, _call_callback_maybe_halt,
                         _wrap_callback, _check_unknown_options,
                         _prepare_scalar_function)
 from ._constraints import old_bound_to_new
