+++ scipy/optimize/_linprog.py
@@ -16,7 +16,8 @@ Functions
 
 import numpy as np
 
-from ._optimize import OptimizeResult, OptimizeWarning
+from scipy._lib._util import OptimizeResult
+from ._optimize import OptimizeWarning
 from warnings import warn
 from ._linprog_highs import _linprog_highs
 from ._linprog_ip import _linprog_ip
