+++ scipy/optimize/_linprog_ip.py
@@ -20,10 +20,11 @@ References
 
 import numpy as np
 import scipy as sp
+from scipy._lib._util import OptimizeResult
 import scipy.sparse as sps
 from warnings import warn
 from scipy.linalg import LinAlgError
-from ._optimize import OptimizeWarning, OptimizeResult, _check_unknown_options
+from ._optimize import OptimizeWarning, _check_unknown_options
 from ._linprog_util import _postsolve
 has_umfpack = True
 has_cholmod = True
