+++ scipy/optimize/_linprog_simplex.py
@@ -30,7 +30,9 @@ References
 
 import numpy as np
 from warnings import warn
-from ._optimize import OptimizeResult, OptimizeWarning, _check_unknown_options
+
+from scipy._lib._util import OptimizeResult
+from ._optimize import OptimizeWarning, _check_unknown_options
 from ._linprog_util import _postsolve
