+++ scipy/optimize/_milp.py
@@ -6,7 +6,7 @@ from numpy.exceptions import VisibleDeprecationWarning
 from scipy.sparse import csc_array, vstack, issparse
 from ._highspy._highs_wrapper import _highs_wrapper  # type: ignore[import-not-found,import-untyped]
 from ._constraints import LinearConstraint, Bounds
-from ._optimize import OptimizeResult
+from scipy._lib._util import OptimizeResult
 from ._linprog_highs import _highs_to_scipy_status_message
