+++ scipy/optimize/_minpack_py.py
@@ -7,9 +7,9 @@ from numpy import (atleast_1d, triu, shape, transpose, zeros, prod, greater,
                    finfo, inexact, issubdtype, dtype)
 from scipy import linalg
 from scipy.linalg import svd, cholesky, solve_triangular, LinAlgError
-from scipy._lib._util import _asarray_validated, _lazywhere, _contains_nan
-from scipy._lib._util import getfullargspec_no_self as _getfullargspec
-from ._optimize import OptimizeResult, _check_unknown_options, OptimizeWarning
+from scipy._lib._util import (OptimizeResult, _asarray_validated, _lazywhere,
+                              _contains_nan, getfullargspec_no_self as _getfullargspec)
+from ._optimize import _check_unknown_options, OptimizeWarning
 from ._lsq import least_squares
 # from ._lsq.common import make_strictly_feasible
 from ._lsq.least_squares import prepare_bounds
