+++ scipy/optimize/_qap.py
@@ -2,10 +2,10 @@ import numpy as np
 import operator
 import warnings
 import numbers
-from . import (linear_sum_assignment, OptimizeResult)
+from . import (linear_sum_assignment)
 from ._optimize import _check_unknown_options
 
-from scipy._lib._util import check_random_state
+from scipy._lib._util import OptimizeResult, check_random_state
 import itertools
 
 QUADRATIC_ASSIGNMENT_METHODS = ['faq', '2opt']
