+++ scipy/optimize/_shgo.py
@@ -8,11 +8,11 @@ import sys
 import numpy as np
 
 from scipy import spatial
-from scipy.optimize import OptimizeResult, minimize, Bounds
+from scipy.optimize import minimize, Bounds
 from scipy.optimize._optimize import MemoizeJac
 from scipy.optimize._constraints import new_bounds_to_old
 from scipy.optimize._minimize import standardize_constraints
-from scipy._lib._util import _FunctionWrapper
+from scipy._lib._util import _FunctionWrapper, OptimizeResult
 
 from scipy.optimize._shgo_lib._complex import Complex
