+++ scipy/optimize/_slsqp_py.py
@@ -16,10 +16,11 @@ Functions
 __all__ = ['approx_jacobian', 'fmin_slsqp']
 
 import numpy as np
+from scipy._lib._util import OptimizeResult
 from scipy.optimize._slsqp import slsqp
 from numpy import (zeros, array, linalg, append, concatenate, finfo,
                    sqrt, vstack, isfinite, atleast_1d)
-from ._optimize import (OptimizeResult, _check_unknown_options,
+from ._optimize import (_check_unknown_options,
                         _prepare_scalar_function, _clip_x_for_func,
                         _check_clip_x)
 from ._numdiff import approx_derivative
