+++ scipy/optimize/_spectral.py
@@ -4,7 +4,7 @@ Spectral Algorithm for Nonlinear Equations
 import collections
 
 import numpy as np
-from scipy.optimize import OptimizeResult
+from scipy._lib._util import OptimizeResult
 from scipy.optimize._optimize import _check_unknown_options
 from ._linesearch import _nonmonotone_line_search_cruz, _nonmonotone_line_search_cheng
