+++ scipy/optimize/_tnc.py
@@ -32,8 +32,9 @@ value of the function, and whose second argument is the gradient of the function
 (as a list of values); or None, to abort the minimization.
 """
 
+from scipy._lib._util import OptimizeResult
 from scipy.optimize import _moduleTNC as moduleTNC
-from ._optimize import (MemoizeJac, OptimizeResult, _check_unknown_options,
+from ._optimize import (MemoizeJac, _check_unknown_options,
                        _prepare_scalar_function)
 from ._constraints import old_bound_to_new
 from scipy._lib._array_api import array_namespace
