+++ scipy/optimize/tests/test_zeros.py
@@ -11,10 +11,9 @@ import numpy as np
 from numpy import finfo, power, nan, isclose, sqrt, exp, sin, cos
 
 from scipy import optimize
-from scipy.optimize import (_zeros_py as zeros, newton, root_scalar,
-                            OptimizeResult)
+from scipy.optimize import (_zeros_py as zeros, newton, root_scalar)
 
-from scipy._lib._util import getfullargspec_no_self as _getfullargspec
+from scipy._lib._util import OptimizeResult, getfullargspec_no_self as _getfullargspec
 
 # Import testing parameters
 from scipy.optimize._tstutils import get_tests, functions as tstutils_functions
